+++ src/Command/ListTubesWatched.php
@@ -11,7 +11,6 @@ namespace Phlib\Beanstalk\Command;
 class ListTubesWatched implements CommandInterface
 {
     use StatsTrait;
-    use ToStringTrait;
 
     public function getCommand(): string
     {