+++ src/Command/Put.php
@@ -15,7 +15,6 @@ use Phlib\Beanstalk\ValidateTrait;
 class Put implements CommandInterface
 {
     use ValidateTrait;
-    use ToStringTrait;
 
     protected string $data;