+++ src/Command/Stats.php
@@ -11,7 +11,6 @@ namespace Phlib\Beanstalk\Command;
 class Stats implements CommandInterface
 {
     use StatsTrait;
-    use ToStringTrait;
 
     public function getCommand(): string
     {