+++ src/Command/Watch.php
@@ -15,7 +15,6 @@ use Phlib\Beanstalk\ValidateTrait;
 class Watch implements CommandInterface
 {
     use ValidateTrait;
-    use ToStringTrait;
 
     protected string $tube;